refactor(events): rename in_memory_event_store to event_store and update imports; streamline event serialization in processes and repository modules ✨
